Ippokratis_kotsanis_131_BC_Assignment_1/Ippokratis_kotsanis_131_BC_Assignment_1/requirements.txt
base58check==1.0.2
bitcoin-utils==0.5.9
ecdsa==0.17.0
mpmath==1.3.0
pycoin==0.92.20230326
python-bitcoinrpc==1.0
six==1.16.0
sympy==1.11.1
